+++ packages/core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Support `unsafeSkipStorageCheck` option in `ValidationOptions`. ([#566](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/566))
+- Include OpenEthereum dev chain as a development network. ([#571](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/571))
 
 ## 1.14.2 (2022-04-18)